LayoutTests/fast/canvas/canvas-path-object-expected.txt

Ensure that the constructor for Path object and dependent functions exist.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof document.getElementById is "function"
PASS path is an instance of Path
PASS typeof path.closePath is "function"
PASS typeof path.moveTo is "function"
PASS typeof path.lineTo is "function"
PASS typeof path.quadraticCurveTo is "function"
PASS typeof path.bezierCurveTo is "function"
PASS typeof path.arcTo is "function"
PASS typeof path.arc is "function"
PASS typeof path.rect is "function"
PASS successfullyParsed is true

TEST COMPLETE
